+++ tests/src/test/scala-3/com/softwaremill/macwire/AutowireCompileTests.scala
@@ -5,7 +5,10 @@ class AutowireCompileTests extends CompileTestsSupport:
 
   runTestsWith(
     expectedFailures = List(
-      "privateConstructor" -> List("Cannot find a dependency or a public constructor for: A.", "Wiring path: A"),
+      "privateConstructor" -> List(
+        "Cannot find a provided dependency, public constructor or public apply method for: A.",
+        "Wiring path: A"
+      ),
       "cyclicDependency" -> List("Cyclic dependencies detected.", "Wiring path: A -> B -> A"),
       "primitives" -> List("Cannot use a primitive type or String in autowiring.", "Wiring path: A -> String")
     )
